+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod abi;
 pub mod aot_generator;
+pub mod code_builder;
 pub mod context;
 pub mod dummy;
 pub mod wavm;
