+++ tests/test_bugs.rs
@@ -1,5 +1,6 @@
 use wasc::abi;
 use wasc::aot_generator;
+use wasc::code_builder;
 use wasc::context;
 use wasc::dummy;
 use wasc::wavm;
@@ -23,14 +24,17 @@ fn test_spec_single_test<P: AsRef<std::path::Path>>(
     abi::init(&mut middle)?;
 
     dummy::init(&mut middle)?;
-    let mut dummy_file = dummy::CodeBuilder::open(&middle.dummy)?;
+    let mut dummy_file = code_builder::CodeBuilder::open(&middle.dummy)?;
     dummy_file.write_line(format!("#include \"{}_glue.h\"", middle.file_stem).as_str())?;
     dummy_file.write_line(
         format!("#include \"./{}_abi/spectest.h\"", middle.file_stem.clone()).as_str(),
     )?;
+    dummy_file.write_line("")?;
     dummy_file.write_line("int main() {")?;
+    dummy_file.intend();
     dummy_file.write_line("wavm_ret_int32_t wavm_ret = wavm_exported_function_main(NULL);")?;
     dummy_file.write_line("return wavm_ret.value;")?;
+    dummy_file.extend();
     dummy_file.write_line("}")?;
     dummy::gcc_build(&middle)?;
